added first prototype of restart functionality
it's now possible to define restarts in any function
using the method "restart". syntax-wise, it's in par
of the "handle" function. if a restart is defined,
you're able to invoke a restart using the function "invoke(restart_name)".

though it's possible to invoke restarts without
binding a condition to a handler, it's pretty useless.
common practice is to invoke a restart in a conditon handler.

since you (mostly) don't want to rewind the stack to the point where
the handler is defined, use bind instead of handle.