removed :NoHandlerFound Condition
removed because the error was caught by the wrong handler
resulting in a unexpected behaviour. this is only caused
by the bootstrapping and will not occur in normal usage
(well, at least i dont was able to reproduce this behaviour
under normal conditions).

also changed policy regarding unsetting conditionhandlers:
condition handlers set by bind() (without block) will not be
unset during execution of the handler. handle, and bind with block
will unset the condition handler as expected.

please keep this in mind.